Add kivy GUI.
Uses kivy to create a GUI for playing dominoes, which can be found
in play.py. Also adds tile_images dir for images of each of the
dominoes.

The GUI is still WIP. It only shows the main row right now, and
the dominoes are just lined up without any orientation. There is a
TODO comment in play.py with next steps.

Minor (untested) changes to the basic classes as well:
* Knock/draw now represented as None, None
* Game.make_move returns success bool
* Game now has make_move_or_knock method
* Player exposes function for parsing input