Add tile index
Tried to speed up get valid moves by keeping track of ends and
remaining tiles in index, but seems to be slower now. Need to add
tests to the board to be confident that this is even working (though
games seem to play successfully). Need to debug why it's slower.